feat: create RegisterController for only sign up
